update ruby (3.4.7) and bundler

+++ db/seeds.rb
@@ -1,3 +1,4 @@
+return if Rails.env.test? || Rails.env.production?
 
 Transfer.destroy_all
 Affiliation.destroy_all
@@ -24,28 +25,28 @@ greens_nsw = Group.create(name: 'The Greens NSW')
 greens_federal = Group.create(name: 'The Greens Federal')
 greens_vic = Group.create(name: 'The Greens Victoria')
 
-Membership.create(group: alp, person: john, title: 'Member', start_date: 2.years.ago)
-Membership.create(group: alp, person: eddie, title: 'Member', start_date: 25.years.ago, end_date: 10.years.ago)
+Membership.create(group: alp, member: john, start_date: 2.years.ago)
+Membership.create(group: alp, member: eddie, start_date: 25.years.ago, end_date: 10.years.ago)
 
-Membership.create(group: usyd, person: john, title: 'Student', start_date: 34.years.ago, end_date: 30.years.ago)
-Membership.create(group: usyd, person: mark, title: 'Student', start_date: 34.years.ago, end_date: 30.years.ago)
-Membership.create(group: usyd, person: ben, title: 'Student', start_date: 34.years.ago, end_date: 30.years.ago)
-Membership.create(group: usyd, person: fox, title: 'Student', start_date: 34.years.ago, end_date: 30.years.ago)
+Membership.create(group: usyd, member: john, start_date: 34.years.ago, end_date: 30.years.ago)
+Membership.create(group: usyd, member: mark, start_date: 34.years.ago, end_date: 30.years.ago)
+Membership.create(group: usyd, member: ben, start_date: 34.years.ago, end_date: 30.years.ago)
+Membership.create(group: usyd, member: fox, start_date: 34.years.ago, end_date: 30.years.ago)
 
-Membership.create(group: aloyisus, person: paul, title: 'Student', start_date: 40.years.ago, end_date: 35.years.ago)
-Membership.create(group: aloyisus, person: ben, title: 'Student', start_date: 40.years.ago, end_date: 35.years.ago)
-Membership.create(group: aloyisus, person: fox, title: 'Student', start_date: 40.years.ago, end_date: 35.years.ago)
+Membership.create(group: aloyisus, member: paul, start_date: 40.years.ago, end_date: 35.years.ago)
+Membership.create(group: aloyisus, member: ben, start_date: 40.years.ago, end_date: 35.years.ago)
+Membership.create(group: aloyisus, member: fox, start_date: 40.years.ago, end_date: 35.years.ago)
 
-Membership.create(group: balance, person: paul, title: 'Owner', start_date: 16.years.ago)
-Membership.create(group: kennards, person: fox, title: 'CIO', start_date: 6.years.ago)
+Membership.create(group: balance, member: paul, start_date: 16.years.ago)
+Membership.create(group: kennards, member: fox, start_date: 6.years.ago)
 
-Membership.create(group: greens_nsw, person: richard, title: 'Member', start_date: 12.years.ago, end_date: 10.years.ago)
-Membership.create(group: greens_vic, person: charles, title: 'Member', start_date: 12.years.ago)
+Membership.create(group: greens_nsw, member: richard, start_date: 12.years.ago, end_date: 10.years.ago)
+Membership.create(group: greens_vic, member: charles, start_date: 12.years.ago)
 
 Transfer.create(giver: balance, taker: greens_nsw, amount: 1000, effective_date: 2.years.ago)
 Transfer.create(giver: kennards, taker: greens_nsw, amount: 1000, effective_date: 2.years.ago)
 Transfer.create(giver: paul, taker: alp, amount: 1000, effective_date: 20.years.ago)
 
-Affiliation.create(owning_group: greens_federal, sub_group: greens_nsw, start_date: 20.years.ago)
-Affiliation.create(owning_group: greens_federal, sub_group: greens_vic, start_date: 20.years.ago)
+Membership.create(group: greens_federal, member: greens_nsw, start_date: 20.years.ago)
+Membership.create(group: greens_federal, member: greens_vic, start_date: 20.years.ago)
 AdminUser.create!(email: 'admin@example.com', password: 'password', password_confirmation: 'password') if Rails.env.development?
